An Idea
This is just an idea of what it could look like this, but I don't have enough information on the code to make good judgements on how to handle the architecture of where methods should be placed (e.g, track() ). It will probably break when running, so don't accept the pull request but try having a look and see what could be done. Good luck!